+++ frontend/src/components/ChatContainer/ChatContainer.tsx
@@ -1,6 +1,4 @@
-import { useRef } from 'react';
-
-import { useAiChatMessage } from '@business/hooks/useAiChat';
+import type { Message } from '@components/ChatContainer';
 
 import ChatInput from './ChatInput';
 import ChatList from './ChatList';
@@ -9,14 +7,12 @@ interface ChatContainerProps {
   width: string;
   height: string;
   position?: string;
+  messages: Message[];
+  inputDisabled: boolean;
+  onSubmitMessage: (message: string) => void;
 }
 
-function ChatContainer({ width, height, position }: ChatContainerProps) {
-  const tarotCardId = useRef<number>();
-
-  // TODO: 현재 무조건 AI 채팅으로 넘어감 -> useHumanChatMessage로 변경 필요
-  const { messages, inputDisabled, onSubmitMessage } = useAiChatMessage(tarotCardId);
-
+function ChatContainer({ width, height, position, messages, inputDisabled, onSubmitMessage }: ChatContainerProps) {
   return (
     <div className={`${width} ${height} ${position} absolute`}>
       <ChatList messages={messages} />
