+++ frontend/src/components/ChatContainer/index.ts
@@ -1,1 +1,3 @@
 export { default } from './ChatContainer';
+
+export type { Message, MessageButton } from './types';
