+++ frontend/src/pages/AIChatPage/AIChatPage.tsx
@@ -4,7 +4,7 @@ import Background from '@components/Background';
 import ChatContainer from '@components/ChatContainer';
 import Header from '@components/Header';
 
-import { useTarotSpread } from '@business/hooks/useAiChat';
+import { useAiChatMessage, useTarotSpread } from '@business/hooks/useAiChat';
 
 interface AIChatPageProps {}
 
@@ -12,6 +12,7 @@ function AIChatPage({}: AIChatPageProps) {
   const tarotCardId = useRef<number>();
 
   useTarotSpread(tarotCardId);
+  const { messages, inputDisabled, onSubmitMessage } = useAiChatMessage(tarotCardId);
 
   return (
     <Background type="dynamic">
@@ -21,6 +22,9 @@ function AIChatPage({}: AIChatPageProps) {
         width="w-700"
         height="h-3/4"
         position="top-90"
+        messages={messages}
+        inputDisabled={inputDisabled}
+        onSubmitMessage={onSubmitMessage}
       />
     </Background>
   );
